requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fonttools==4.58.5
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
mangum==0.19.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.1
pillow==11.2.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydyf==0.11.0
Pygments==2.19.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-whois==0.9.5
pytz==2025.2
PyYAML==6.0.2
pyzbar==0.1.9
qrcode==8.1
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
rich-toolkit==0.14.3
scikit-learn==1.7.0
scipy==1.16.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tldextract==5.3.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
zopfli==0.2.3.post1
